// src/types/mock-data.type.ts
export type MockData = {
  name: string[];
  description: string[];
  // publictionDate:
  released: number[];
  previewVideoImage: string[];
  previewVideoLink: string[];
  starring: string[];
  director: string[];
  // runTime: number;
  // commentCount: number;
  posterImage: string[];
  backgroundImage: string[];
  backgroundColor: string[];
  userName: string[];
  userEmail: string[];
  userAvatar: string[];
};